requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
mysql-connector-python==8.2.0
transformers==4.36.2
torch==2.1.2
langchain==0.0.350
langchain-huggingface==0.0.1
langchain-community==0.0.10
python-dotenv==1.0.0
